pyproject.toml
[project]
name = "mcp-auth-oidc"
version = "0.1.0"
description = "MCP server with OAuth 2.1 authentication via Auth0"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

